Minor bug fixes and modmail replies allowed.
Minor bug fixes and modmail replies allowed.
#1 epoch time is not 1440 * days its 86400*days.
#2 dont log tickets from automod or reddit itself.
#3 removed a few semi colons
Also added support to allow replies directly from the ticketing system
into modmail.  System exited if it was unable to update the ticket to
remove the custom field data as it could result in a loop that would
cause infinite replies to reddit (a bad thing).