#6 Add feature to reopen resolved tickets.
Three new configuration values added to support automatically reopening
tickets if there is a reply.  The status's that we change from/to are
configurable so you could, if you wanted, reopen deleted tickets.  I
think thats a bad idea - from a workflow standpoint, deleted tickets can
be for things that have no value / worthless (modmail trolling) while
resolved are things that are completed.  The reopen of the ticket is
non-crashing so if it fails due to a request tracker failure, it won't
cause the modmail script to fail but it will log an error to the
console.  Configuration section is getting a bit unwieldy so I may move
this to its own config.py at some point.  Not a high priority.