#8 Added new settings and made reddit posts safer
Updated setting for default sleep interval from 5 seconds to 60 seconds
- reddit has been slow to accept api posts lately even after they sent
an ack - up to ~20 seconds -  so the new feature to try to be careful
would be less than helpful if we were only waiting 5 seconds.  An extra
good bit of time here won't really matter.
Added 3 new settings that are MANDATORY that control the initial ticket
subject, comment, and reply comments.  These are tokenized strings so
you can change these to whatever they are desired but these three things
must 100% be present.
Updated functionality such that if we see a ticket comment to push to
modmail we will make a good faith effort to make sure this hasn't
already been pushed to reddit (and possibly succeeded but reddit timed
out letting us know that it did).  Algorithm explanation in ticket #8.