Updates to fix breaking changes (python 3+ diffs)
Fixed a few syntactic sugar items added to python 3 that were used here
that were not available in older versions of python.  Some shared hosts
still only have python 2.6 installed so still want to support the latest
security patched python there (2.6.8).  Basically, removed the name
parameter from a stream handler as the name changed later on and also
added the string-format indexes back in.

Documented the goal is for python 2.6.8 or higher code and also updated
to include a new dependency on argparse that was added.

Tested fine.