Allow root-authors to be delegated to queues
Specifically, this was added to support the use
case where someone wishes automoderator
created modmail to not be ignored (as we have
been doing) but to be created in a separate queue.
In my example, I have automoderator going to queue 1
this should change depending on which queue you
wish it to go to obviously.  Request tracker changes to
support the new queue are same as always.  Make sure
that the request tracker account has full permissions
to this new queue and make sure that the custom
field for modmail replies is 'applies to' this new queue
if you wish to allow this new queue to be used for
modmail replies as well.  Be aware of the security
and reddit considerations for this if you allow
access to this queue for lower privileged users.